+++ src/engine/wv/rendering/mesh.h
@@ -31,7 +31,7 @@ struct GeometrySurface
 
 	std::vector<Primitive> primitives;
 	
-	std::vector<uint16_t> indices;
+	std::vector<uint32_t> indices;
 
 	std::vector<Vector3f> vertexPositions;
 	std::vector<Vector3f> vertexNormals;
